created RestaurantsScreen component & moved the SafeAreaView from App component to RestaurantScreen component
