+++ src/features/restaurants/components/restaurants-info.component.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import {View,Text,StyleSheet} from 'react-native'
+import RestaurantsScreen from '../screens/restaurants.screens';
+
+
+const RestaurantsInfo=()=>(
+  <Text>RestaurantsInfo</Text>
+)
+
+export default RestaurantsInfo;
